Rearchitected Project
- Added Maven, in order to help build project easier
- Restructured Package structure, in order to conform to Java standards
