--- ServerNaptime/src/plugin.yml
@@ -1,11 +1,0 @@
-main: gvk.ServerNaptimeMain
-name: ServerNaptime
-version: 1.0
-api-version: 1.16
-description: Server takes a nap while no players are online
-author: gvk
-commands:
-  naptime:
-    description: Command that will toggle ServerNaptime.
-    permission: naptime.toggle
-    permission-message: You do not have the required permission (<permission>)
